Attempt to change Windows compilation to cross compile on Ubuntu
I tried to copy the way VVVVVV-CE does it, but I give this 10% chance
of working immediately because everything was made completely
different.
Right now I also don't have a lot of time on my hands to figure out
how to do this any different, or how to not use NixOS because I'm not
Leo.