Include DLLs in Windows package
It's not going to be a drop-in replacement for Windows anymore, since
the DLLs you need depend on the compiler used. I only didn't bother
with data.zip yet.

I could probably also have changed the original VS to a non-debug build
or something in the first place, but I've gone though all this trouble,
and I don't know how I would do that (I don't see an existing debug
switch or whatever in the VS command). It's probably easier for me to
just add in the needed DLLs for a cross-compiled build, than to revert
everything and find out where to start on that. Plus, it's probably
more convenient for the user to already have the DLLs there anyway.