Add start of (basically proof-of-concept) localization system
This might still have too much STL, but it more or less works easily
and understandably now.

This adds Localization.cpp and Localization.h, which have a loc
namespace. Basically, a translation file in XML format can be read in,
and loc::gettext("Game paused") in place of "Game paused" looks up the
translation based on the English string, gettext-style. (Where needed,
there could be a cgettext function that has an extra context specifier)
One string ("Game paused") is wrapped in loc::gettext as a demo.

At the very least, there will also need to be support for auto-wrapping
certain strings that are now split manually (like in menu option
descriptions) and roomnames and cutscenes need more specialized ways,
too.