Add automatic word wrapping
Right now only the description for analogue mode is automatically
wordwrapped as a demo. The wordwrapping algorithm, Graphics::wordwrap,
is fairly simple, it splits a string based on spaces, and returns a new
string with possibly some strings replaced by '\n'.

Graphics::PrintWrap is the automatically wrapping print function, to be
used for text that is currently manually wrapped and split over multiple
calls to Graphics::Print.