Remove (start of) plural forms system
I like the concept of the plural form system, but I think we can
probably simplify it away here. There are not many places where it'd be
necessary, only place I can think of right now is when you collect a
crewmate, and it distinguishes between "One remains" or "Two remain",
which could be changed to "One remaining" or "Two remaining".
Meanwhile, there is also a place where the existing English doesn't
care ("You rescued One crewmates and found One trinkets.") and while it
would be nice to improve that, it may not be worth the extra code
complexity and the fact that plural forms alone mean you can't add a
language without hardcoding something in the C++ (the plural expression
for that language). Either way, we still have the code in case I change
my mind.