Change language file XML structure a bit
The `comment` attribute is now called `explanation`. There's now a new
`comment` field, empty by default, intended for translators to enter
their comments (because any added XML comments would be removed when
the language files are synced), though keeping these comments is not
yet implemented.

I also documented a more elaborate `max` attribute, which would also
give the maximum number of columns and lines in wrapped text.

I made the changes to the structure with TinyXML itself, but I didn't
commit the code for that (I just inverted the check for non-English
language files in the sync function and made the attribute changes
there). This changed the English strings.xml to spaces indentation
instead of tabs indentation, so I just decided to make all the other
language-related XMLs (basically only meta.xmls) consistently indented
with spaces.

While documenting the maximum string lengths and knowing I barely give
those for the menus, I realized menus are really a problem for
localization because their positioning is hardcoded, yet hard to
specify, and _even manually checking all menus_ isn't enough because
options could suddenly be added in the middles of menus (like in
PR #313) moving previously fitting options further to the right. So I
think I'll try to fix that in a separate PR for a bit by making menus
automatically centered and spaced based on the options.