Replace "wrapped" notes by column and row count in language files
The max attribute can now be in the format X*Y, which means, maximum
X columns by Y rows, auto-wrapped (unless disabled for that language).
Sometimes these are a bit of an estimate, but I did check many of them
by measurement, as far as I know there are no maximum indications
that are too high.