Allow "Untitled Level" and "By Unknown" to be translated
At first I thought it wouldn't be a good idea to make the default level
title and author translatable, because the translations would just end
up being posted online as part of level files in English-speaking
communities - many people whose native language is not English don't
restrict themselves to visiting communities in their own language. Tons
of levels already exist that have these two strings as well. So a
better idea is to treat the title "Untitled Level" and the creator
"Unknown" as special; they will be displayed in the user's language,
and even sorted in the levels list accordingly.

In the level editor, the title and creator are also rendered as the
translation if they're set to the default English one, and they still
get saved in the level file in English. Pressing the button to change
the title or creator will however carry over that translation and make
it editable.