Allow editor message to have multiple lines
The message banner that comes up in the middle of the screen when
changing the tileset, saving a level, etc is now wordwrapped and can
thus have more than a screen wide of text. This will make it easier for
translators, especially when they have to translate "Now using %s
Tileset" and all the tileset names like "Space Station" separately.