Add support for translating the textual numbers
Originally I said the UtilityClass::number overhaul would need to be
reverted for localization, but I had a better idea: since the game
should fully function in English if language files are completely
missing, that function can stay the way it is for English anyway (we
never refer to an English language file to know the numbers), and refer
to a translated full list if the language is not English.

Which brings me to the full list of numbers. Different languages
assemble numbers differently, so you can forget about sticking Twenty
and Four together to get 24 in any language. The only real option is to
let translators translate all numbers, and the list goes to 100
nowadays. This list is numbers.xml, which also works separately from
the other strings, since we can just use numbers as a key.

By the way, there are apparently also languages which don't represent
numbers the same way in all cases where VVVVVV currently assumes they
are. For example, it was noted that "Twenty out of Twenty" would need
to be translated as "dwadzieścia z dwudziestu" in Polish. Therefore,
translations for the numbers can also be left blank, in which case
numeric representation will be used automatically (20 out of 20).