Add "balanced" wordwrapping function for dialogue
This function will attempt to find a the best textbox width for a given
dialogue, where all lines are closest to each other in length. This
will be used in translated cutscenes, because instead of a textbox like
this:

| Something's wrong! We're going to |
| crash!                            |

...you'd want the textbox to look more like this:

| Something's wrong! We're |
| going to crash!          |

It just calls the wordwrap function multiple times to try out for which
width (between the given bounds, for cutscene dialogues I'm thinking
probably between 16-32 characters per line) you get the lowest
percentage of post-line whitespace, and returns the best one.