Add most of the support for cutscene translation
All main-game cutscenes can now be translated! cutscenes.xml for each
language contains all the (unwrapped) pieces of dialogue in a logical
order, along with the speaker/colour for each line for context.

Example:

<cutscene id="firststeps" explanation="">
    <dialogue speaker="cyan" english="I wonder why the ship teleported
        me here alone?">Waarom zou het schip mij hier in mijn eentje
        hebben geteleporteerd?</dialogue>
    <dialogue speaker="cyan" english="I hope everyone else got out
        ok...">Ik hoop dat de rest in orde is...</dialogue>
</cutscene>

The idea is that the original cutscenes themselves remain unchanged
where possible, and translators still don't need to worry about text
wrapping (except for CJK). There ARE a few dialogues where this won't
work, like the dimensional stability generator terminal, these will
need some special handling and will probably need to be manually
wrapped.

I've already translated the first few cutscenes to Dutch and it's all
working nicely thus far.

Support for custom levels' cutscenes can later use mostly the same
system with little changes, because in the end they also use the text
command and have distinct script names. The only difference is that a
dedicated cutscenes.xml would need to be loaded when the level is
loaded (somewhere like lang/XX/customlevels/LEVELNAME/cutscenes.xml?
This would tie in really nicely with the per-level custom assets
feature as well!)

On a technical note, interestingly scriptclass::scriptname was in
place, but was completely unused (it was only set to "null" in
scriptclass::hardreset()). Now it's actually used for its (probably)
intended purpose, to store the name of the currently running script.