Add support for translating the specialline textboxes
These are the "I'm worried about (CREWMATE), Doctor!" and "He's really
looking forward to helping you find (CREWMATE/the rest of the crew)!"
lines. The original string concatenation code stays for English, and
every combination is translated separately as regular dialogue in
cutscenes.xml. There are not that many cases and different languages
most definitely have different grammar depending on who is addressed
and which of the two sentences it is.

And the "you" cases are real. I'm not fixing that :D
https://distractionware.com/forum/index.php?topic=1116.msg14317#msg14317